+++ src/main/java/com/foodforfriends/web/controller/ReviewController.java
@@ -1,4 +1,4 @@
-package com.foodforfriends.web.review;
+package com.foodforfriends.web.controller;
 
 import java.net.URISyntaxException;
 import java.util.Collection;
@@ -14,6 +14,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.foodforfriends.model.Review;
+import com.foodforfriends.web.service.ReviewService;
 
 @RestController
 public class ReviewController {
